refactor(observability): gather the settings helpers in one file
Review: the copied deserializer does not belong to the alarm catalogue. It moves to
`settings/utils.rs` alongside `validate_config_ids`, which every section already shares, so the
sections hold configuration and the helpers sit beside each other.
